refactor(hooks): split os.rs into os/ module with per-tool files
